fix review handler, components
