+++ src/components/incident-report/ServiceInfo.js
@@ -1,0 +1,10 @@
+export default function ServiceInfo() {
+  return (
+    <div className="bg-primary-dark rounded-lg sm:rounded-xl p-3 sm:p-4 md:p-6 lg:p-8 text-white">
+      <h2 className="text-sm sm:text-base md:text-lg lg:text-xl font-bold mb-2 sm:mb-3 md:mb-4 lg:mb-6">Apa Itu Layanan Aduan Siber?</h2>
+      <p className="text-xs sm:text-sm md:text-base leading-relaxed text-gray-200">
+        Layanan Aduan Siber adalah sistem pelaporan insiden keamanan siber yang memungkinkan pengguna untuk melaporkan berbagai ancaman seperti serangan <span className="text-primary-teal font-medium">phishing</span>, <span className="text-primary-teal font-medium">pembebaran malware</span>, atau <span className="text-primary-teal font-medium">penyalahgunaan data</span>. Tim CSIRT akan menganalisis laporan dan memberikan solusi sesuai dengan tingkat urgensi insiden.
+      </p>
+    </div>
+  );
+}
