+++ src/components/incident-report/VulnerabilityTypes.js
@@ -1,0 +1,28 @@
+export default function VulnerabilityTypes() {
+  const vulnerabilities = [
+    'Cross-site Scripting (XSS)',
+    'Cross-site Request Forgery',
+    'Server-Side Request Forgery (SSRF)',
+    'SQL Injection',
+    'Server-side Remote Code Execution (RCE)',
+    'XML External Entity Attacks (XXE)',
+    'Access Control Issues (Insecure Direct Object Reference issues, etc.)',
+    'Exposed Administrative Panels that don\'t require login credentials'
+  ];
+
+  return (
+    <div className="bg-primary-dark rounded-lg sm:rounded-xl p-3 sm:p-4 md:p-6 lg:p-8 text-white">
+      <h2 className="text-sm sm:text-base md:text-lg lg:text-xl font-bold mb-2 sm:mb-3 md:mb-4 lg:mb-6">Jenis Kerentanan</h2>
+      <div className="space-y-2 sm:space-y-3 lg:space-y-4">
+        {vulnerabilities.map((vulnerability, index) => (
+          <div key={index} className="flex items-start">
+            <span className="inline-flex items-center justify-center w-5 h-5 sm:w-6 sm:h-6 md:w-7 md:h-7 bg-primary-teal text-primary-dark text-xs sm:text-sm font-bold rounded mr-2 sm:mr-3 flex-shrink-0 mt-0.5">
+              {index + 1}
+            </span>
+            <span className="text-xs sm:text-sm md:text-base text-gray-200 leading-relaxed">{vulnerability}</span>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+}
